+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Initial clone records script
 * Initial set resource IDs script
 * Initial update collections script
+* Initial set citation script
 
 ### Changed